Used previously present class as base class for OutputFiles and moved it to the io folder, deleted newly developed class.
